HPNet-Argoverse/datamodules/historical_prediction_datamodule.py
import torch
from typing import Optional, List
from torch_geometric.data import Batch

from datamodules.argoverse_v1_datamodule import ArgoverseV1DataModule
from datasets.argoverse_v1_dataset import ArgoverseV1Dataset

class HistoricalPredictionDataModule(ArgoverseV1DataModule):
    """
    增强版数据模块，支持历史预测轨迹生成
    结合 RealMotion 和 HPNet 的套娃思路
    """
    
    def __init__(self, 
                 historical_window: int = 3,  # 历史预测窗口大小
                 root: str = None,
                 train_batch_size: int = None,
                 val_batch_size: int = None,
                 shuffle: bool = True,
                 num_workers: int = 0,
                 pin_memory: bool = True,
                 persistent_workers: bool = True,
                 flip_p: float = 0.5,
                 agent_occlusion_ratio: float = 0.05,
                 lane_occlusion_ratio: float = 0.2,
                 num_historical_steps: int = 20,
                 num_future_steps: int = 30,
                 margin: float = 50,
                 **kwargs):
        super().__init__(
            root=root,
            train_batch_size=train_batch_size,
            val_batch_size=val_batch_size,
            shuffle=shuffle,
            num_workers=num_workers,
            pin_memory=pin_memory,
            persistent_workers=persistent_workers,
            flip_p=flip_p,
            agent_occlusion_ratio=agent_occlusion_ratio,
            lane_occlusion_ratio=lane_occlusion_ratio,
            num_historical_steps=num_historical_steps,
            num_future_steps=num_future_steps,
            margin=margin,
            **kwargs
        )
        self.historical_window = historical_window
        self.historical_predictions_cache = {}  # 缓存历史预测结果
        
    def setup(self, stage: Optional[str] = None):
        super().setup(stage)
        
    def generate_historical_predictions(self, scenario_id: str, current_step: int) -> Optional[torch.Tensor]:
        """
        生成历史预测轨迹
        scenario_id: 场景ID
        current_step: 当前时间步
        """
        if scenario_id not in self.historical_predictions_cache:
            return None
            
        # 获取历史预测轨迹
        historical_preds = self.historical_predictions_cache[scenario_id]
        if current_step < self.historical_window:
            return None
            
        # 提取历史窗口内的预测
        start_step = max(0, current_step - self.historical_window)
        end_step = current_step
        return historical_preds[start_step:end_step]
        
    def update_historical_predictions(self, scenario_id: str, step: int, predictions: torch.Tensor):
        """
        更新历史预测轨迹缓存
        scenario_id: 场景ID
        step: 时间步
        predictions: 预测结果 [(N1,...,Nb), K, F, 2]
        """
        if scenario_id not in self.historical_predictions_cache:
            self.historical_predictions_cache[scenario_id] = {}
            
        self.historical_predictions_cache[scenario_id][step] = predictions
        
    def train_dataloader(self):
        """
        重写训练数据加载器，添加历史预测轨迹处理
        """
        dataloader = super().train_dataloader()
        # 这里可以添加历史预测轨迹的处理逻辑
        # 由于 DataLoader 的 collate_fn 比较复杂，我们暂时在模型层面处理
        return dataloader
        
    def val_dataloader(self):
        """
        重写验证数据加载器，添加历史预测轨迹处理
        """
        dataloader = super().val_dataloader()
        # 这里可以添加历史预测轨迹的处理逻辑
        return dataloader
        
    def on_train_batch_end(self, outputs, batch, batch_idx: int):
        """
        训练批次结束后更新历史预测轨迹
        """
        # 这里可以添加逻辑来更新历史预测轨迹
        # 由于 PyTorch Lightning 的回调机制，这里暂时不实现
        # 实际的历史预测轨迹生成将在模型层面处理
        pass
        
    def on_validation_batch_end(self, outputs, batch, batch_idx: int):
        """
        验证批次结束后更新历史预测轨迹
        """
        pass
        
    def on_test_batch_end(self, outputs, batch, batch_idx: int):
        """
        测试批次结束后更新历史预测轨迹
        """
        pass 