HPNet-Argoverse/requirements.txt
ray
tqdm
hydra-core==1.2.0
setuptools==59.5.0
protobuf==3.19.0
rich==12.4.4
wandb==0.13.5
plotly
matplotlib==3.4.3
timm>=0.6.7
torchmetrics==0.10.2
torch-geometric==2.3.1
mamba-ssm
